// src/interfaces/SearchSubjectOpenEnrResponseInterface.ts
export interface SearchSubjectOpenEnrResponseInterface {
  subjectCode: string;
  subjectNameTh: string;
  subjectNameEn: string;
  credit: string;
  theoryHour: string;
  practiceHour: string;
  selfHour: string;
  subjectType: string;
  flagCur: string;
  creditShow: string;
  relateSubjectCode: string;
}